refactoring: Review of class design.
