+++ src/Fonts/lgfx_fonts.cpp
@@ -1,5 +1,9 @@
 #include "lgfx_fonts.hpp"
 
+#include "../lgfx/lgfx_common.hpp"
+#include "../lgfx/LGFXBase.hpp"
+
+#include <algorithm>
 #include <cstdint>
 #include <cstddef>
 
@@ -28,6 +32,184 @@ namespace lgfx {
     return uniCode < 256;
   }
 
+  std::size_t GLCDfont::drawChar(LGFXBase* me, std::int32_t x, std::int32_t y, std::uint16_t c, const TextStyle* style) const
+  {
+    if (c > 255) return 0;
+
+    if (!style->cp437 && (c >= 176)) c++; // Handle 'classic' charset behavior
+
+    const std::int32_t fontWidth  = this->width;
+    const std::int32_t fontHeight = this->height;
+
+    auto font_addr = this->chartbl + (c * 5);
+    std::uint32_t colortbl[2] = {me->getColorConverter()->convert(style->back_rgb888), me->getColorConverter()->convert(style->fore_rgb888)};
+    bool fillbg = (style->back_rgb888 != style->fore_rgb888);
+
+    //std::int32_t clip_left   = me->_clip_l;
+    //std::int32_t clip_right  = me->_clip_r;
+    //std::int32_t clip_top    = me->_clip_t;
+    //std::int32_t clip_bottom = me->_clip_b;
+
+    float sy = style->size_y;
+    float sx = style->size_x;
+    //if ((x <= clip_right) && (clip_left < (x + fontWidth * sx ))
+    // && (y <= clip_bottom) && (clip_top < (y + fontHeight * sy )))
+    {
+//      if (!fillbg || style->size_y != 1.0 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * sx > clip_right) {
+        std::int32_t x1 = sx;
+        std::int32_t x0 = 0;
+        me->startWrite();
+
+        std::int_fast8_t i = 0;
+        do {
+          std::uint8_t line = font_addr[i];
+          std::uint8_t flg = (line & 0x01);
+          std::int_fast8_t j = 1;
+          std::int_fast16_t y0 = 0;
+          std::int_fast16_t y1 = 0;
+          do {
+            while (flg == ((line >> j) & 0x01) && ++j < fontHeight);
+            y1 = j * sy;
+            if (flg || fillbg) {
+              me->setRawColor(colortbl[flg]);
+              me->writeFillRect(x + x0, y + y0, x1 - x0, y1 - y0);
+            }
+            y0 = y1;
+            flg = !flg;
+          } while (j < fontHeight);
+          x0 = x1;
+          x1 = (++i + 1) * sx;
+        } while (i < fontWidth - 1);
+
+        if (fillbg) {
+          me->setRawColor(colortbl[0]);
+          me->writeFillRect(x + x0, y, x1 - x0, fontHeight * style->size_y); 
+        }
+        me->endWrite();
+/*
+      } else {
+        std::uint8_t col[fontWidth];
+        std::int_fast8_t i = 0;
+        do {
+          col[i] = font_addr[i];
+        } while (++i < 5);
+        col[5] = 0;
+        me->startWrite();
+        me->setAddrWindow(x, y, fontWidth * style->size_x, fontHeight);
+        std::uint8_t flg = col[0] & 1;
+        std::uint32_t len = 0;
+        i = 0;
+        do {
+          std::int_fast8_t j = 0;
+          do {
+            if (flg != ((col[j] >> i) & 1)) {
+              me->writeRawColor(colortbl[flg], len);
+              len = 0;
+              flg = !flg;
+            }
+            len += style->size_x;
+          } while (++j < fontWidth);
+        } while (++i < fontHeight);
+        me->writeRawColor(colortbl[0], len);
+        me->endWrite();
+      }
+//*/
+    }
+    return fontWidth * sx;
+  }
+
+  static std::size_t draw_char_bmp(LGFXBase* me, std::int32_t x, std::int32_t y, const TextStyle* style, const std::uint8_t* font_addr, std::int_fast8_t fontWidth, std::int_fast8_t fontHeight, std::int_fast8_t w, std::int_fast8_t margin )
+  {
+    std::uint32_t colortbl[2] = {me->getColorConverter()->convert(style->back_rgb888), me->getColorConverter()->convert(style->fore_rgb888)};
+    bool fillbg = (style->back_rgb888 != style->fore_rgb888);
+
+    //std::int32_t clip_left   = me->_clip_l;
+    //std::int32_t clip_right  = me->_clip_r;
+    //std::int32_t clip_top    = me->_clip_t;
+    //std::int32_t clip_bottom = me->_clip_b;
+
+    float sx = style->size_x;
+    std::int32_t sh = fontHeight * style->size_y;
+
+    //if ((x <= clip_right) && (clip_left < (x + fontWidth * sx ))
+    // && (y <= clip_bottom) && (clip_top < (y + sh )))
+    {
+//      if (!fillbg || sy != 1 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * sx > clip_right) {
+        me->startWrite();
+        if (fillbg) {
+          me->setRawColor(colortbl[0]);
+          if (margin) {
+            std::int32_t x0 = (fontWidth - margin) * sx;
+            std::int32_t x1 = (fontWidth         ) * sx;
+            if (x0 < x1) {
+              me->writeFillRect(x + x0, y, x1 - x0, sh);
+            }
+          }
+        }
+        std::int32_t i = 0;
+        std::int32_t y1 = 0;
+        std::int32_t y0 = - 1;
+        do {
+          bool fill = y0 != y1;
+          y0 = y1;
+          y1 = ++i * sh / fontHeight;
+          std::uint8_t line = font_addr[0];
+          bool flg = line & 0x80;
+          std::int_fast8_t j = 1;
+          std::int_fast8_t je = fontWidth - margin;
+          std::int32_t x0 = 0;
+          do {
+            do {
+              if (0 == (j & 7)) line = font_addr[j >> 3];
+            } while (flg == (bool)(line & (0x80) >> (j&7)) && ++j < je);
+            std::int32_t x1 = j * sx;
+            if (flg || (fillbg && fill)) {
+              me->setRawColor(colortbl[flg]);
+              if (flg && x1 == std::int32_t((j-1)*sx)) ++x1;
+              me->writeFillRect(x + x0, y + y0, x1 - x0, std::max<std::int32_t>(1, y1 - y0));
+            }
+            x0 = x1;
+            flg = !flg;
+          } while (j < je);
+          font_addr += w;
+        } while (i < fontHeight);
+        me->endWrite();
+/*
+      } else {
+        std::int_fast8_t len = 0;
+        std::uint8_t line = 0;
+        bool flg = false;
+        me->startWrite();
+        me->setAddrWindow(x, y, fontWidth * style->size_x, fontHeight);
+        std::int_fast8_t i = 0;
+        std::int_fast8_t je = fontWidth - margin;
+        do {
+          std::int_fast8_t j = 0;
+          do {
+            if (j & 7) {
+              line <<= 1;
+            } else {
+              line = (j == je) ? 0 : font_addr[j >> 3];
+            }
+            if (flg != (bool)(line & 0x80)) {
+              me->writeRawColor(colortbl[flg], len);
+              flg = !flg;
+              len = 0;
+            }
+            len += style->size_x;
+          } while (++j < fontWidth);
+          font_addr += w;
+        } while (++i < fontHeight);
+        me->writeRawColor(colortbl[flg], len);
+        me->endWrite();
+      }
+//*/
+    }
+
+    return fontWidth * sx;
+  }
+
+
   bool BMPfont::updateFontMetric(FontMetrics *metrics, std::uint16_t uniCode) const {
     if ((uniCode -= 32) >= 96) return false;
     metrics->x_advance = metrics->width = widthtbl[uniCode];
@@ -37,6 +219,106 @@ namespace lgfx {
   bool BDFfont::updateFontMetric(FontMetrics *metrics, std::uint16_t uniCode) const {
     metrics->x_advance = metrics->width = (uniCode < 0x0100) ? halfwidth : width;
     return true;
+  }
+
+  std::size_t BMPfont::drawChar(LGFXBase* me, std::int32_t x, std::int32_t y, std::uint16_t uniCode, const TextStyle* style) const
+  { // BMP font
+    if ((uniCode -= 0x20) >= 96) return 0;
+    const std::int_fast8_t fontWidth = this->widthtbl[uniCode];
+    const std::int_fast8_t fontHeight = this->height;
+
+    auto font_addr = ((const std::uint8_t**)this->chartbl)[uniCode];
+    return draw_char_bmp(me, x, y, style, font_addr, fontWidth, fontHeight, (fontWidth + 6) >> 3, 1);
+  }
+
+  std::size_t BDFfont::drawChar(LGFXBase* me, std::int32_t x, std::int32_t y, std::uint16_t c, const TextStyle* style) const
+  {
+    const std::int_fast8_t bytesize = (this->width + 7) >> 3;
+    const std::int_fast8_t fontHeight = this->height;
+    const std::int_fast8_t fontWidth = (c < 0x0100) ? this->halfwidth : this->width;
+    auto it = std::lower_bound(this->indextbl, &this->indextbl[this->indexsize], c);
+    if (*it != c) {
+      if (style->fore_rgb888 != style->back_rgb888) {
+        me->fillRect(x, y, fontWidth * style->size_x, fontHeight * style->size_y, style->back_rgb888);
+      }
+      return fontWidth * style->size_x;
+    }
+    const std::uint8_t* font_addr = &this->chartbl[std::distance(this->indextbl, it) * fontHeight * bytesize];
+    return draw_char_bmp(me, x, y, style, font_addr, fontWidth, fontHeight, bytesize, 0);
+  }
+
+  std::size_t RLEfont::drawChar(LGFXBase* me, std::int32_t x, std::int32_t y, std::uint16_t code, const TextStyle* style) const
+  { // RLE font
+    if ((code -= 0x20) >= 96) return 0;
+
+    const int fontWidth = this->widthtbl[code];
+    const int fontHeight = this->height;
+
+    auto font_addr = ((const std::uint8_t**)this->chartbl)[code];
+
+    std::uint32_t colortbl[2] = {me->getColorConverter()->convert(style->back_rgb888), me->getColorConverter()->convert(style->fore_rgb888)};
+    bool fillbg = (style->back_rgb888 != style->fore_rgb888);
+
+    //std::int32_t clip_left   = me->_clip_l;
+    //std::int32_t clip_right  = me->_clip_r;
+    //std::int32_t clip_top    = me->_clip_t;
+    //std::int32_t clip_bottom = me->_clip_b;
+
+    float sx = style->size_x;
+    float sy = style->size_y;
+
+    //if ((x <= clip_right) && (clip_left < (x + fontWidth  * sx ))
+    // && (y <= clip_bottom) && (clip_top < (y + fontHeight * sy )))
+    {
+//      if (!fillbg || sy != 1.0 || x < clip_left || y < clip_top || y + fontHeight > clip_bottom || x + fontWidth * sx > clip_right) {
+        bool flg = false;
+        std::uint8_t line = 0, i = 0, j = 0;
+        std::int32_t len;
+        std::int32_t y0 = 0;
+        std::int32_t y1 = sy;
+        std::int32_t x0 = 0;
+        me->startWrite();
+        do {
+          line = *font_addr++;
+          flg = line & 0x80;
+          line = (line & 0x7F)+1;
+          do {
+            len = (j + line > fontWidth) ? fontWidth - j : line;
+            line -= len;
+            j += len;
+            std::int32_t x1 = j * sx;
+            if (fillbg || flg) {
+              me->setRawColor(colortbl[flg]);
+              me->writeFillRect( x + x0, y + y0, x1 - x0, y1 - y0);
+            }
+            x0 = x1;
+            if (j == fontWidth) {
+              j = 0;
+              x0 = 0;
+              y0 = y1;
+              y1 = (++i + 1) * sy;
+            }
+          } while (line);
+        } while (i < fontHeight);
+        me->endWrite();
+/*
+      } else {
+        std::uint32_t line = 0;
+        me->startWrite();
+        me->setAddrWindow(x, y, fontWidth * style->size_x, fontHeight);
+        std::uint32_t len = fontWidth * style->size_x * fontHeight;
+        do {
+          line = *font_addr++;
+          bool flg = line & 0x80;
+          line = ((line & 0x7F) + 1) * style->size_x;
+          me->writeRawColor(colortbl[flg], line);
+        } while (len -= line);
+        me->endWrite();
+      }
+//*/
+    }
+
+    return fontWidth * sx;
   }
 
 
@@ -99,6 +381,446 @@ namespace lgfx {
     metrics->height   = glyph_bb + glyph_ab;
     metrics->y_advance = yAdvance;
   }
+
+//----------------------------------------------------------------------------
+
+  void VLWfont::getDefaultMetric(FontMetrics *metrics) const {
+    metrics->x_offset  = 0;
+    metrics->y_offset  = 0;
+    metrics->baseline  = maxAscent;
+    metrics->y_advance = yAdvance;
+    metrics->height    = yAdvance;
+  }
+
+  VLWfont::~VLWfont() {
+    unloadFont();
+  }
+
+  bool VLWfont::unloadFont(void) {
+    _fontLoaded = false;
+    if (gUnicode)  { heap_free(gUnicode);  gUnicode  = nullptr; }
+    if (gWidth)    { heap_free(gWidth);    gWidth    = nullptr; }
+    if (gxAdvance) { heap_free(gxAdvance); gxAdvance = nullptr; }
+    if (gdX)       { heap_free(gdX);       gdX       = nullptr; }
+    if (gBitmap)   { heap_free(gBitmap);   gBitmap   = nullptr; }
+    if (_fontData) {
+      _fontData->preRead();
+      _fontData->close();
+      _fontData->postRead();
+      _fontData = nullptr;
+    }
+    return true;
+  }
+
+  bool VLWfont::getUnicodeIndex(std::uint16_t unicode, std::uint16_t *index) const
+  {
+    auto poi = std::lower_bound(gUnicode, &gUnicode[gCount], unicode);
+    *index = std::distance(gUnicode, poi);
+    return (*poi == unicode);
+  }
+
+  bool VLWfont::updateFontMetric(FontMetrics *metrics, std::uint16_t uniCode) const {
+    std::uint16_t gNum = 0;
+    if (getUnicodeIndex(uniCode, &gNum)) {
+      if (gWidth && gxAdvance && gdX[gNum]) {
+        metrics->width     = gWidth[gNum];
+        metrics->x_advance = gxAdvance[gNum];
+        metrics->x_offset  = gdX[gNum];
+      } else {
+        auto file = _fontData;
+
+        file->preRead();
+
+        file->seek(28 + gNum * 28);  // headerPtr
+        std::uint32_t buffer[6];
+        file->read((std::uint8_t*)buffer, 24);
+        metrics->width    = __builtin_bswap32(buffer[1]); // Width of glyph
+        metrics->x_advance = __builtin_bswap32(buffer[2]); // xAdvance - to move x cursor
+        metrics->x_offset  = (std::int32_t)((std::int8_t)__builtin_bswap32(buffer[4])); // x delta from cursor
+
+        file->postRead();
+      }
+      return true;
+    }
+    if (uniCode == 0x20) {
+      metrics->width = metrics->x_advance = metrics->y_advance * 2 / 7;
+      metrics->x_offset = 0;
+      return true;
+    }
+    return false;
+  }
+
+
+  bool VLWfont::loadFont(DataWrapper* data) {
+    _fontData = data;
+    {
+      std::uint32_t buf[6];
+      data->read((std::uint8_t*)buf, 6 * 4); // 24 Byte read
+
+      gCount   = __builtin_bswap32(buf[0]); // glyph count in file
+                //__builtin_bswap32(buf[1]); // vlw encoder version - discard
+      yAdvance = __builtin_bswap32(buf[2]); // Font size in points, not pixels
+                //__builtin_bswap32(buf[3]); // discard
+      ascent   = __builtin_bswap32(buf[4]); // top of "d"
+      descent  = __builtin_bswap32(buf[5]); // bottom of "p"
+    }
+
+    // These next gFont values might be updated when the Metrics are fetched
+    maxAscent  = ascent;   // Determined from metrics
+    maxDescent = descent;  // Determined from metrics
+    yAdvance   = std::max((int)yAdvance, ascent + descent);
+    spaceWidth = yAdvance * 2 / 7;  // Guess at space width
+
+//ESP_LOGI("LGFX", "ascent:%d  descent:%d", gFont.ascent, gFont.descent);
+
+    if (!gCount) return false;
+
+//ESP_LOGI("LGFX", "font count:%d", gCount);
+
+    std::uint32_t bitmapPtr = 24 + (std::uint32_t)gCount * 28;
+
+    gBitmap   = (std::uint32_t*)heap_alloc_psram( gCount * 4); // seek pointer to glyph bitmap in the file
+    gUnicode  = (std::uint16_t*)heap_alloc_psram( gCount * 2); // Unicode 16 bit Basic Multilingual Plane (0-FFFF)
+    gWidth    =  (std::uint8_t*)heap_alloc_psram( gCount );    // Width of glyph
+    gxAdvance =  (std::uint8_t*)heap_alloc_psram( gCount );    // xAdvance - to move x cursor
+    gdX       =   (std::int8_t*)heap_alloc_psram( gCount );    // offset for bitmap left edge relative to cursor X
+
+    if (nullptr == gBitmap  ) gBitmap   = (std::uint32_t*)heap_alloc( gCount * 4); // seek pointer to glyph bitmap in the file
+    if (nullptr == gUnicode ) gUnicode  = (std::uint16_t*)heap_alloc( gCount * 2); // Unicode 16 bit Basic Multilingual Plane (0-FFFF)
+    if (nullptr == gWidth   ) gWidth    =  (std::uint8_t*)heap_alloc( gCount );    // Width of glyph
+    if (nullptr == gxAdvance) gxAdvance =  (std::uint8_t*)heap_alloc( gCount );    // xAdvance - to move x cursor
+    if (nullptr == gdX      ) gdX       =   (std::int8_t*)heap_alloc( gCount );    // offset for bitmap left edge relative to cursor X
+
+    if (!gUnicode
+      || !gBitmap
+      || !gWidth
+      || !gxAdvance
+      || !gdX) {
+//ESP_LOGE("LGFX", "can not alloc font table");
+      return false;
+    }
+
+    _fontLoaded = true;
+
+    size_t gNum = 0;
+    _fontData->seek(24);  // headerPtr
+    std::uint32_t buffer[7];
+    do {
+      _fontData->read((std::uint8_t*)buffer, 7 * 4); // 28 Byte read
+      std::uint16_t unicode = __builtin_bswap32(buffer[0]); // Unicode code point value
+      std::uint32_t w = (std::uint8_t)__builtin_bswap32(buffer[2]); // Width of glyph
+      if (gUnicode)   gUnicode[gNum]  = unicode;
+      if (gWidth)     gWidth[gNum]    = w;
+      if (gxAdvance)  gxAdvance[gNum] = (std::uint8_t)__builtin_bswap32(buffer[3]); // xAdvance - to move x cursor
+      if (gdX)        gdX[gNum]       =  (std::int8_t)__builtin_bswap32(buffer[5]); // x delta from cursor
+
+      std::uint16_t height = __builtin_bswap32(buffer[1]); // Height of glyph
+      if ((unicode > 0xFF) || ((unicode > 0x20) && (unicode < 0xA0) && (unicode != 0x7F))) {
+        std::int16_t dY =  (std::int16_t)__builtin_bswap32(buffer[4]); // y delta from baseline
+//Serial.printf("LGFX:unicode:%x  dY:%d\r\n", unicode, dY);
+        if (maxAscent < dY && unicode != 0x3000) {
+          maxAscent = dY;
+        }
+        if (maxDescent < (height - dY) && unicode != 0x3000) {
+//Serial.printf("LGFX:maxDescent:%d\r\n", maxDescent);
+          maxDescent = height - dY;
+        }
+      }
+
+      if (gBitmap)  gBitmap[gNum] = bitmapPtr;
+      bitmapPtr += w * height;
+    } while (++gNum < gCount);
+
+    yAdvance = maxAscent + maxDescent;
+
+//Serial.printf("LGFX:maxDescent:%d\r\n", maxDescent);
+    return true;
+  }
+
+//----------------------------------------------------------------------------
+
+  std::size_t GFXfont::drawChar(LGFXBase* me, std::int32_t x, std::int32_t y, std::uint16_t uniCode, const TextStyle* style) const
+  {
+    auto glyph = this->getGlyph(uniCode);
+    if (!glyph) return 0;
+
+    std::int32_t w = glyph->width,
+                  h = glyph->height;
+
+    float sx = style->size_x;
+    float sy = style->size_y;
+
+    std::int32_t xAdvance = sx * glyph->xAdvance;
+    std::int32_t xoffset  = sx * glyph->xOffset;
+
+    me->startWrite();
+    std::uint32_t colortbl[2] = {me->getColorConverter()->convert(style->back_rgb888), me->getColorConverter()->convert(style->fore_rgb888)};
+    bool fillbg = (style->back_rgb888 != style->fore_rgb888);
+    std::int32_t left  = 0;
+    std::int32_t right = 0;
+    if (fillbg) {
+      left  = std::max<int>(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
+      right = x + std::max<int>(w * sx + xoffset, xAdvance);
+      me->setRawColor(colortbl[0]);
+    }
+
+    x += xoffset;
+    y += int(me->_font_metrics.y_offset * sy);
+    std::int32_t yoffset = (- me->_font_metrics.y_offset) + glyph->yOffset;
+
+    //std::int32_t clip_left   = me->_clip_l;
+    //std::int32_t clip_right  = me->_clip_r;
+    //std::int32_t clip_top    = me->_clip_t;
+    //std::int32_t clip_bottom = me->_clip_b;
+
+    //if ((x <= clip_right) && (clip_left < (x + w * sx ))
+    // && (y <= clip_bottom) && (clip_top < (y + h * sy )))
+    {
+
+      if (left < right) {
+        if (yoffset > 0) {
+          me->writeFillRect(left, y, right - left, yoffset * sy);
+        }
+        std::int32_t y0 = (yoffset + h) * sy;
+        std::int32_t y1 = me->_font_metrics.height * sy;
+        if (y0 < y1) {
+          me->writeFillRect(left, y + y0, right - left, y1 - y0);
+        }
+      }
+
+      std::uint8_t *bitmap = &this->bitmap[glyph->bitmapOffset];
+      std::uint8_t mask=0x80;
+
+      me->setRawColor(colortbl[1]);
+      std::int_fast8_t i = 0;
+      std::int32_t y1 = yoffset * sy;
+      std::int32_t y0 = y1 - 1;
+      do {
+        bool fill = y0 != y1;
+        y0 = y1;
+        y1 = (++i + yoffset) * sy;
+        std::int32_t fh = y1 - y0;
+        if (!fh) fh = 1;
+        if (left < right && fill) {
+          me->setRawColor(colortbl[0]);
+          me->writeFillRect(left, y + y0, right - left, fh);
+          me->setRawColor(colortbl[1]);
+        }
+
+        std::int32_t j = 0;
+        std::int32_t x0 = 0;
+        bool flg = false;
+        do {
+          do {
+            if (flg != (bool)(*bitmap & mask)) break;
+            if (! (mask >>= 1)) {
+              mask = 0x80;
+              ++bitmap;
+            }
+          } while (++j < w);
+          std::int32_t x1 = j * sx;
+          if (flg) {
+            std::int32_t fw = (x0 < x1) ? x1 - x0 : 1;
+            me->writeFillRect(x + x0, y + y0, fw, fh);
+          }
+          x0 = x1;
+          flg = !flg;
+        } while (j < w);
+      } while (i < h);
+    //} else {
+    //  if (left < right) {
+    //    me->writeFillRect(left, y, right - left, (me->_font_metrics.height) * sy);
+    //  }
+    }
+    me->_filled_x = right;
+    me->endWrite();
+    return xAdvance;
+  }
+
+  std::size_t VLWfont::drawChar(LGFXBase* lgfxbase, std::int32_t x, std::int32_t y, std::uint16_t code, const TextStyle* style) const
+  {
+    auto me = (LGFXBase*)lgfxbase;
+    auto file = this->_fontData;
+
+    std::uint32_t buffer[6] = {0};
+    std::uint16_t gNum = 0;
+
+    if (code == 0x20) {
+      gNum = 0xFFFF;
+      buffer[2] = __builtin_bswap32(this->spaceWidth);
+    } else if (!this->getUnicodeIndex(code, &gNum)) {
+      return 0;
+    } else {
+      file->preRead();
+      file->seek(28 + gNum * 28);
+      file->read((std::uint8_t*)buffer, 24);
+      file->seek(this->gBitmap[gNum]);
+    }
+
+
+    std::int32_t h        = __builtin_bswap32(buffer[0]); // Height of glyph
+    std::int32_t w        = __builtin_bswap32(buffer[1]); // Width of glyph
+    float sx = style->size_x;
+    std::int32_t xAdvance = __builtin_bswap32(buffer[2]) * sx; // xAdvance - to move x cursor
+    std::int32_t xoffset   = (std::int32_t)((std::int8_t)__builtin_bswap32(buffer[4])) * sx; // x delta from cursor
+    std::int32_t dY        = (std::int16_t)__builtin_bswap32(buffer[3]); // y delta from baseline
+    float sy = style->size_y;
+    std::int32_t yoffset = (this->maxAscent - dY);
+//      std::int32_t yoffset = (me->_font_metrics.y_offset) - dY;
+
+    std::uint8_t pbuffer[w * h];
+    std::uint8_t* pixel = pbuffer;
+    if (gNum != 0xFFFF) {
+      file->read(pixel, w * h);
+      file->postRead();
+    }
+
+    me->startWrite();
+
+    std::uint32_t colortbl[2] = {me->getColorConverter()->convert(style->back_rgb888), me->getColorConverter()->convert(style->fore_rgb888)};
+    bool fillbg = (style->back_rgb888 != style->fore_rgb888);
+    std::int32_t left  = 0;
+    std::int32_t right = 0;
+    if (fillbg) {
+      left  = std::max(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
+      right = x + std::max<int>(w * sx + xoffset, xAdvance);
+    }
+    me->_filled_x = right;
+
+    x += xoffset;
+    y += int(me->_font_metrics.y_offset * sy);
+
+    std::int32_t bx = x;
+    std::int32_t bw = w * sx;
+    std::int32_t clip_left = me->_clip_l;
+    if (x < clip_left) { bw += (x - clip_left); bx = clip_left; }
+
+    std::int32_t clip_right = me->_clip_r;
+    if (bw > clip_right+1 - bx) bw = clip_right+1 - bx;
+
+    if (bw >= 0)
+    {
+      std::int32_t fore_r = ((style->fore_rgb888>>16)&0xFF);
+      std::int32_t fore_g = ((style->fore_rgb888>> 8)&0xFF);
+      std::int32_t fore_b = ((style->fore_rgb888)    &0xFF);
+
+      if (fillbg || !me->isReadable() || me->hasPalette())
+      { // fill background mode  or unreadable panel  or palette sprite mode
+        if (left < right && fillbg) {
+          me->setRawColor(colortbl[0]);
+          if (yoffset > 0) {
+            me->writeFillRect(left, y, right - left, yoffset * sy);
+          }
+          std::int32_t y0 = (yoffset + h) * sy;
+          std::int32_t y1 = me->_font_metrics.height * sy;
+          if (y0 < y1) {
+            me->writeFillRect(left, y + y0, right - left, y1 - y0);
+          }
+        }
+
+        if (0 < w) {
+          uint32_t back = fillbg ? style->back_rgb888 : me->_base_rgb888;
+          std::int32_t back_r = ((back>>16)&0xFF);
+          std::int32_t back_g = ((back>> 8)&0xFF);
+          std::int32_t back_b = ( back     &0xFF);
+          std::int32_t i = 0;
+          std::int32_t y0, y1 = yoffset * sy;
+          do {
+            y0 = y1;
+            if (y0 > me->_clip_b) break;
+            y1 = (yoffset + i + 1) * sy;
+            if (left < right) {
+              me->setRawColor(colortbl[0]);
+              me->writeFillRect(left, y + y0, right - left, y1 - y0);
+            }
+            std::int32_t j = 0;
+            do {
+              std::int32_t x0 = j * sx;
+              while (pixel[j] != 0xFF) {
+                std::int32_t x1 =(j+1)* sx;
+                if (pixel[j] != 0 && x0 < x1) {
+                  std::int32_t p = 1 + (std::uint32_t)pixel[j];
+                  me->setColor(color888( ( fore_r * p + back_r * (257 - p)) >> 8
+                                        , ( fore_g * p + back_g * (257 - p)) >> 8
+                                        , ( fore_b * p + back_b * (257 - p)) >> 8 ));
+                  me->writeFillRect(x + x0, y + y0, x1 - x0, y1 - y0);
+                }
+                x0 = x1;
+                if (++j == w || clip_right < x0) break;
+              }
+              if (j == w || clip_right < x0) break;
+              me->setRawColor(colortbl[1]);
+              do { ++j; } while (j != w && pixel[j] == 0xFF);
+              me->writeFillRect(x + x0, y + y0, (j * sx) - x0, y1 - y0);
+            } while (j != w);
+            pixel += w;
+          } while (++i < h);
+        }
+      }
+      else // alpha blend mode
+      {
+        bgr888_t buf[bw * (int)ceil(sy)];
+        pixelcopy_t p(buf, me->getColorConverter()->depth, rgb888_3Byte, me->hasPalette());
+        std::int32_t y0, y1 = yoffset * sy;
+        std::int32_t i = 0;
+        do {
+          y0 = y1;
+          if (y0 > me->_clip_b) break;
+          y1 = (yoffset + i + 1) * sy;
+          std::int32_t by = y + y0;
+          std::int32_t bh = y1 - y0;
+
+          auto ct = me->_clip_t;
+          if (by < ct) { bh += by - ct; by = ct; }
+          if (bh > 0) {
+            std::int32_t j0 = 0;
+            std::int32_t j1 = w;
+
+            // search first and last pixel
+            while (j0 != j1 && !pixel[j0    ]) { ++j0; }
+            while (j0 != j1 && !pixel[j1 - 1]) { --j1; }
+
+            if (j0 != j1) {
+              std::int32_t rx = j0  * sx;
+              std::int32_t rw = j1 * sx;
+              if (rx < bx    -x) rx = bx    -x;
+              if (rw > bx+bw -x) rw = bx+bw -x;
+              rw -= rx;
+
+              if (0 < rw) {
+                me->readRectRGB(x + rx, by, rw, bh, (std::uint8_t*)buf);
+
+                std::int32_t x0, x1 = (j0 * sx) - rx;
+                do {
+                  x0 = x1;
+                  if (x0 < 0) x0 = 0;
+                  x1 = (int)((j0+1) * sx) - rx;
+                  if (x1 > rw) x1 = rw;
+                  if (pixel[j0] && x0 < x1) {
+                    std::int32_t p = 1 + pixel[j0];
+                    do {
+                      std::int32_t yy = 0;
+                      do {
+                        auto bgr = &buf[x0 + yy * rw];
+                        bgr->r = ( fore_r * p + bgr->r * (257 - p)) >> 8;
+                        bgr->g = ( fore_g * p + bgr->g * (257 - p)) >> 8;
+                        bgr->b = ( fore_b * p + bgr->b * (257 - p)) >> 8;
+                      } while (++yy != bh);
+                    } while (++x0 != x1);
+                  }
+                } while (++j0 < j1);
+                me->pushImage(x + rx, by, rw, bh, &p);
+              }
+            }
+          }
+          pixel += w;
+        } while (++i < h);
+      }
+    }
+    me->endWrite();
+    return xAdvance;
+  }
+
 }
 
 //----------------------------------------------------------------------------
